software/fixture_io/requirements.txt
pyserial==3.4
pytest==5.4.1
